Two Sum
Intuition
The first thing I remembered was arrays. For me, a vector is an advanced array. Knowing the principle of working with the elements of a regular array, it was easy to do the same with a vector.

Approach
We need to find the indices of two vector elements, the sum of these elements must equal the target. Using two cycles for we compare the sum of zero(In C++ the counting of elements in an array starts with 0) and the first element with the target, in case of false we iterate our indexes and compare the next elements.

Complexity
Time complexity: O(n^2)

Space complexity: O(1)